python_ai/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
requests==2.31.0
pandas==2.1.4
openpyxl==3.1.2
pydantic==2.5.3
python-dotenv==1.0.0
pymysql
matplotlib

python-multipart
numpy
opencv-python
langchain>=0.1.0
langchain-core
langchain-openai